Add TLSv1.1 and TLSv1.2 in mod_ssl.pl
This patch adds TLSv1.1 and TLSv1.2 to @sslprotos in mod_ssl.pl.

Without this patch the user will have to manually enable TLSv1.2
support in their apache configuration. TLSv1.2 support is required
to recieve a good SSL score on the Qualys SSL test (otherwise the
score is capped to B).