Add support for replacements in help link via URL
* Can be useful for replacements like WordPress with ClassicPress or MySQL with MariaDB

Usage: 5th parameter must be is an array reference of hash references, e.g.:

[ { 'WordPress' => 'ClassicPress' } ]

It is useful when `help_pre_load` cannot be used during a normal call, like in WP Workbench when has to be called via underlying API (e.g. Virtualmin script's API) and cannot be used during the module call.